chore: Removal of unnecessary function, and smooth improvement of good practices

+++ src/models/Products.ts
@@ -27,7 +27,7 @@ export const Products = db.define(
       allowNull: false,
     },
     oldPrice: {
-      type: INTEGER,
+      type: DECIMAL,
       allowNull: true,
     },
     price: {
